Start volunteering platform

+++ database/migrations/2017_06_19_150219_add_banned_at_column_to_users_table.php
@@ -27,8 +27,10 @@ class AddBannedAtColumnToUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('banned_at');
-        });
+        if (! Schema::hasTable('users')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('banned_at');
+            });
+        }
     }
 }
